Split into library and a test client
