+++ .gitignore
@@ -2,3 +2,5 @@
 node_modules
 downloaded
 clientKey.txt
+lib/clientKey.txt
+.env
